fix bounding box computation, use geopy for distance computation
