+++ requirements.txt
@@ -1,7 +1,8 @@
-reportlab~=4.0.5
+reportlab~=4.0.8
 svglib~=1.5.1
-Pillow~=10.0.1
+Pillow~=10.2.0
 pdf2image~=1.16.3
 requests~=2.31.0
-geopandas~=0.14.0
-shapely~=2.0.1
+geopandas~=0.14.1
+shapely~=2.0.2
+geopy~=2.4.1
